Change behaviour of equals button
The button was causing a bug when used in a way that wasn't initially
expected.

If the equals button was pressed after entering only one number and one
operator, and then the user tried to use it properly, entering the
number that should follow the operator, the operation was ignored by
the calculator, even though now it could actually be done.